+++ Sources/csuper-gui/main.c
@@ -60,6 +60,8 @@ int main (int   argc, char *argv[])
     data.ptr_clipboard=gtk_clipboard_get(gdk_atom_intern("CLIPBOARD",TRUE));
     data.ptr_clipboard_selected=gtk_clipboard_get(gdk_atom_intern("PRIMARY",TRUE));
     data.ptr_csu_struct_tmp=NULL;
+    data.indexLastCsuStruct=0;
+    data.nbLastCsuStruct=0;
 
     filename =  g_build_filename("csuper-gui.glade", NULL);
 
@@ -80,6 +82,7 @@ int main (int   argc, char *argv[])
 
     noCsuFileRanking(&data);
     noCsuFilePoints(&data);
+    setButtonMainWindowSensitive(&data);
 
     gtk_widget_show_all(data.ptr_main_window);